search songの実装

+++ domain/src/main/java/com/snowdango/bijouk/domain/api/entity/ContentVersion.kt
@@ -10,5 +10,5 @@ data class ContentVersion(
     @SerialName("MZ_INDEXER")
     val mzIndexer: Long,
     @SerialName("RTCI")
-    val rtci: Int
+    val rtci: Long,
 )
